Support passing full `requirements.txt` to the `packages` option
The full `requirements.txt` format supports markers. They allow
conditionally depending on packages based on platform or Python version.
For example, the following line in a `requirements.txt` file will only
install the dependency on Windows:
```
pywin32==300; sys_platform == "win32"
```
This is a crucial feature to enable multi-platform support.

As a nice extra benefit, this simplifies how we pass the `packages`
specification to Conan. We no longer need `.split(“\n”)`. See the
test package for example.