Add support for Linux and macOS
This is an MVP implementation of Linux and macOS support. Everything is
functional but there are two caveats that should certainly be resolved
in future versions:
 * The standard library is not pre-compiled and zipped so it takes up
   more space than the Windows variant.
 * The environment is not as locked down as the Windows variant:
   `pip` is still accessible in the final package.

On Windows, we can simply download a pre-built binary embedded package,
but on macOS and Linux, we need to build from source and lock down the
environment ourselves. This means that the two points above require a
bit more effort. We can tackle this later. For now, this implementation
is certainly good enough to get going.