Fix slow `find_package(Python)` on Windows
It seems that, even though we give it an exact `Python_ROOT_DIR`,
CMake does still search the system for all other Python versions.
In the end, it does correctly select the one in `Python_ROOT_DIR`
as we specified but the this entire process is very slow: between
20 seconds and a full minute.

The fix is to give CMake more info: tell it which version of Python
is at `Python_ROOT_DIR` and request it to be EXACT. This seems to
allow it to return immediately without considering others. It brings
the time down to ~3 seconds. And if we also tell it the exact
location of `Python_EXECUTABLE`, the time is down to ~1 second.