Fix `packages` option breaking on comments in `requirements.txt`
Reading in a raw `requirements.txt` file and passing it to the
`packages` option would produce an error if comments are present
in the file.

Our `packages` option is of type `ANY` which Conan takes in as any
string. The problem is that Conan doesn’t like newline characters in
that string (it uses newlines internally for serialization to delimit
individual options). Newline separators worked accidentally for us
up until now (quirk in the Conan’s (de)serialization logic). But it
definitely breaks if a `requirements.txt` file includes comments.

To work around this limitation, we also allow `packages` to be a
tab-separated list. This is a bit weird and reading in a `requirements.txt`
file requires `.replace(“\n”, “\t”)` as you can see from the test
package. But I don’t see a nicer solution and this does certainly work.