Fix build with the latest Python 3.9 and 3.10 versions
We used the web installer getting headers and `python3.lib`. It looks
like it's been removed for the latest versions in favor of the full
offline installer or individual `.msi` files. Downloading only the
specific `dev.msi` file directly is even easier than getting it via
the web installer, so let's do that instead.