Fix build with `cmake_find_package` generator
The issue was that the CMake variables `CONAN_EMBEDDED_PYTHON_ROOT` and
`CONAN_USER_EMBEDDED_PYTHON_pyversion` are only set when using the `cmake`
generator but not with `cmake_find_package`.

The solution is to avoid using these variables and instead rely on the
fact that `Python_ROOT_DIR` is in the same directory as the `embedded_python.cmake`
file. And for the version number, we do a string replacement in `conanfile.py`.
It's fine to hardcode this since the version is always unique to the
Conan package.