Fixed build parameters so dependency zlib is used
After updating to conan 2.0 AutoToolsDeps doesn't add include paths `-I` to `CPPFLAGS` as AutotoolsBuildEnvironment does so when Python was being build and `zlib` was required to unzip packages, it will only look for it in the system folders.

This commit also reformat the recipe so it complies more with Conan 2.0 style.